move lecture fields
