+++ graphql/index.ts
@@ -32,7 +32,7 @@ import { MutateTutoringInterestConfirmationResolver } from './tutoring_interest_
 import { MutateParticipationCertificateResolver } from './certificate/mutations';
 import { ExtendedFieldsParticipationCertificateResolver } from './certificate/fields';
 import { ExtendFieldsStudentResolver } from './student/fields';
-import { ExtendedFieldsLectureResolver } from './lecture/fields';
+import { ExtendedFieldsLectureResolver } from './appointment/fields';
 import { MutateMeResolver } from './me/mutation';
 import responseCachePlugin from 'apollo-server-plugin-response-cache';
 import { cacheModelEnhancementMap } from './cache';
